+++ Dockerfile
@@ -32,7 +32,7 @@ FROM build as runtime
 #COPY --from=build ${WORKDIR}/app app/
 #COPY --from=build ${WORKDIR}/config config/
 #COPY --from=build ${WORKDIR}/public public/
-#COPY --from=build ${WORKDIR}/server.mjs ${WORKDIR}/app.js ${WORKDIR}/git.properties.json ./
+#COPY --from=build ${WORKDIR}/server.mjs ${WORKDIR}/app.mjs ${WORKDIR}/git.properties.json ./
 EXPOSE 3000
 CMD ["yarn", "start" ]
